requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.1
atpublic==4.0
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.5
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
dpath==2.1.6
dulwich==0.21.6
dvc==3.29.0
dvc-data==2.20.0
dvc-http==2.30.2
dvc-objects==1.2.0
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
entrypoints==0.4
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.40
grandalf==0.8
gto==1.5.0
hydra-core==1.3.2
idna==3.4
importlib-resources==6.1.1
iterative-telemetry==0.0.8
Jinja2==3.1.2
kombu==5.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
omegaconf==2.3.0
orjson==3.9.10
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
prompt-toolkit==3.0.41
psutil==5.9.6
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydot==1.4.2
pygit2==1.13.2
Pygments==2.16.1
pygtrie==2.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scmrepo==1.4.1
semver==3.0.2
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.1
sqltrie==0.8.0
sympy==1.12
tabulate==0.9.0
tomlkit==0.12.3
torch==2.1.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
voluptuous==0.14.0
wcwidth==0.2.10
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0

numpy~=1.24.4
pandas~=2.0.3
mlflow~=2.9.0